Fixed int issue with windows

+++ run_big.py
@@ -7,7 +7,7 @@ window_width = 1440
 rows = 9
 cols = 16
 window_ratio = cols / rows
-window_height = window_width / window_ratio
+window_height = int(window_width / window_ratio)
 
 # Set up a window
 game_window = pyglet.window.Window(window_width, window_height)
